Fixes for thousand separators and radix points in input, radix points in output
Before, '.' and ',' were always treated as radix points in input. Now
',' is only treated as a radix point if the user's locale specifies it.

If it doesn't, it means you can paste something like "1,574.78".

Before, all numbers were displayed with '.' as their radix point. Now
they respect the user's locale.

The lexer was also reformatted to have roughly 1 order of magnitude fewer
states.