Let the window handle window-specific actions
Pull copy/cut/paste/undo/redo and mode changing out of application scope
and into window scope. Remove mode switching from the app menu and put
it into a gear menu instead. The gear menu is on the left instead of the
right because this looks far better in Basic Mode.

https://bugzilla.gnome.org/show_bug.cgi?id=714991