Use MPC for complex numbers
MPC is a library for complex numbers built on top of MPFR. Using it
fixes a number of issues with GNOME Calculators incomplete support for
complex numbers and simplifies the implementation of a number of
functions.

Thanks to Daniel Renninghoff this commit also reworks the MPFR bindings
to use vala's automatic memory management. I've added a default rounding
mode which makes the calling code cleaner.

The use of MPFR.RealRef is not ideal but there doesn't seem to be a
better way to deal with references to structs in vala.

The lvalue_allowed = false annotation is necessary to make MPC.abs() and
MPC.arg() work, otherwise vala stores the result in a copy of the target
variable.

The bindings for MPC and MPFR.RealRef are licensed GPL3+ as MPFR and
MPC are LGPL3 so it doesn't make sense to license them as GPL2+ as the
GPL2 is incompatible with LGPL3

TODO:
* Fix possible integer overflows when converting for int64 to
long (these are inherited from the previous implementation)
* Add a class property to make complex numbers optional - at the moment
the tests that prevent arguments that would product complex results from
functions like cosh are still in place so as not to lose the
translations.
* Add tests for complex numbers to test-number.vala - at the moment
arg() is the only function tested with complex inputs.
* Improve the behaviour of root and power with fraction real powers so
that the simplest root (i.e. real or imaginary rather than complex) is
returned for things like ∛i which should return -i but currently
returns ½√3+½i