Fix reusing results from the history stack
Fix a regression introduced by 9bb6936ba74602ec891c1ffecdf1665dba1a1be4
(Migrate HistoryEntry to use a GtkBuilder template)

GtkLabel widgets are window-less, so they can't receive the
"button-press-event" signal directly: they need to be wrapped inside
a GtkEventBox. The old procedural code was doing that but in the
transition to the GtkBuilder template this detail was forgotten.

Fix that by just putting a GtkEventBox as a parent of a GtkLabel in
data/history-entry.ui

While at it also remove the deprecated GtkLabel attributes "xalign" and
"yalign", and don't specify the "events" property explicitly since
that's not needed anymore.

https://bugzilla.gnome.org/show_bug.cgi?id=766155