Fixed conversion unit switch, fixes #748732.
https://bugzilla.gnome.org/show_bug.cgi?id=748732
On populating the target combobox, the current value of the from combo
was removed intentionally. This way, setting the old from value as new
target value failed, as it did not exist in the target combo. Note that
this has only happened with the initially selected from currency, so
removing that from target currencies also led to inconsistencies between
the two comboboxes. We better have the same list in both combos, in the
worst case a user can use the same unit as both from and to unit to get
a valid 1 conversion result.