Remember window position
Save the last window position before quitting the application.
Then restore the first window to this position the next time
the application starts. When quitting the application (via
the appmenu), if there are multiple open windows, the active
window position will be saved.

This will allow users to always start the application in a
predefined position on the screen. The feature is especially
useful when multiple application are configured to open
at set locations so that they can be viewed without
overlapping each other.

Note that only the first window position is restored. Subsequent
windows will not be positioned by the application.