Update parser to allow complex conversion.
This patch rewrites the part of parser that handles conversions.

With the new code, it's possible to convert variables, as well as
complex equations.
With this patch, the following are possible.
10×20 m² in acres
1000+500 GB in GiB
((10 inch in cm) × (1 feet in cm)) cm² in m²
2π radians in degrees
3π + π÷2 radians in degrees

This is perticularly useful for people working with multiple units at
the same time.
e.g. Calculating Gravitational force between a 0.25 pound bag of chips
and 1.25kg laptop that are 17 feet away from each other.

https://bugzilla.gnome.org/show_bug.cgi?id=770344