search-provider: Use async calls, cancel search on inactivity
Move to async methods everywhere and factorize the `solve` calls to a single
method that only uses Subprocess and that can be cancelled.

As per this, if the the search-provider is trying to compute some complex
operation, the daemon won't hang and once the applications' inactivity
timeout is hit, any running instance of gnome-calculator will be killed and
the daemon will return accordingly.

This reduces the impact of an issue that can cause gnome-calculator to keep
running forever if a complex computation is required (10!!!) from the shell,
and won't ever be killed (see GNOME/gnome-shell/#183).

This is also a prerequisite for supporting the search Cancellation that is
going to be available on next version of the Search provider protocol
(as per GNOME/gnome-shell/#436)