search-provider: stop normalizing the equation twice
When an equation is given by the g-s overview, g-c
will solve this equation, where it will be first
normalized and then solved.
This is done by running a subprocess, where g-c
will call g-c in this subprocess with the
"--solve" argument.
With this "--solve" argument, the equation will
also be normalized.
However, when an equation is given by the g-s
overview, it will also be normalized in a check,
that happens before the solve process, to not to
process single numbers given by the g-s overview.
This normalized equation will then be used to
invoke the subprocess to solve the equation,
leading into normalizing the already normalized
equation and therefore resulting into wrong
results when the equation itself contains
decimal numbers.

Solve this issue, by invoking the subprocess with
the unnormalized equation, instead of the
normalized one, to not to normalize the
equation twice.

Closes: https://gitlab.gnome.org/GNOME/gnome-calculator/issues/104