Remove mnemonics blocking mode shortcuts (fixes #138)
Due to gtk#860 mnemonics inside the popupbutton primary menu steal
the shortcuts with Alt, and as such the mode selections do not work
on English locale, where Programming mode selection is stolen by the
_Preferences mnemonic, Advanced mode selection stolen by the _About
mnemonic, and Keyboard mode stolen by the _Keyboard shortcuts mnemonic.
Until I find a reasonable workaround or this gets fixed in gtk, removing
the mnemonics for the problematic items.