Add tooltip for bits in programming mode
For slightly more dyslectic people having to count out the bits is quite
difficult. The lower labels help a little, but not enough. Further more,
adding labels for all the bits just becomes noisy, ugly and also hard to
read.

Instead, simply enable tooltips for all digits, so we can see what bit
relates to what bit number.

Note, that the labels and id's are counting in the opposite direction as
what where expected!

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>