Draft of change to currency converter
This addresses the concerns brought up in #150 -- adding an option
to the preferences menu that changes the way currency names are
displayed in the converter. Options are the name (current default),
the currency code (like USD), or both (United States Dollar USD).

Still needs some polishing but this may work.