Fixed display in preferences
The option was not displaying in preferences because I had
forgotten to add the set_combo_box_from_int for the new combo
box. Also, added (requires restart) because the only sane way
to change the displayed labels for currencies is when the list
is created at startup. Shouldn't be a big deal.

Working on getting it to display for financial mode -- the
conversion dropdown in advanced mode does display correctly. I
think this commit fixes it but I'm having issues with Builder.