Financial mode labels partially working
Only the from part of the currency conversion combo boxes are
displaying the currency names according to the new setting. This
is because the changes are in update_from_model, but there doesn't
appear to be an update_to_model and I need to figure out where
the to model is being created/updated.