Removed requirement for restart when currency format is changed
The MathConverter class now connects to the settings changed
signal, looking for changes to the currency-display-format setting.
If this occurs, it updates the from model and also provokes a
callback to from_combobox_changed_cb() to update the displayed
units. However, because of the if (category != to_category) statement
in that function, the to combobox is not being updated.