Edited combobox_changed_cb so that to model gets updated
However, this has introduced a regression: the to model now gets
updated every time the from box is changed, which is not what we
want. There's a couple options to fix this; copy the code that
updates the to model into the new update_currency_labels() function;
have some sort of boolean value that tells us whether the to model
needs to be updated (only set to true when the currency format
setting is changed, and then cleared after the models are updated)
or adding an else statemet to the if (category != to_category)
that updates the models only when the categories are different?

Pushing this so Robert Roth can take a look @robertroth