ci: Remove distro CI in favor of flatpak
This commit removes our distro-reliant CI in favor of
a flatpak-based CI pipeline. This pipeline will be more
reliable, have up-to-date dependencies (e.g. libadwaita),
and produce a flatpak bundle for each MR and push to master.
It will also make the developer version of Calculator
available on the nightly flatpak repository. This means
it will be easier for users and developers alike
to test Calculator.

See https://gitlab.gnome.org/GNOME/Initiatives/-/issues/1