calculator: Destroy GtkMessageDialog after clicking on Cancel
When there are multiple calculator windows open, quitting
gnome-calculator shows a GtkMessageDialog asking the user if they intend
to close all calculator windows.

The 'Close All' button works as intended, but the 'Cancel' button does
nothing and the dialog remains. The user is left with pressing escape
to dismiss the dialog or click on 'Close All' whether or not they
intended to close all calculator windows.

This fix makes sure the dialog is closed/destroyed after a response is
selected. Clicking on the 'Cancel' button should dismiss the dialog as
expected.

https://gitlab.gnome.org/GNOME/gnome-calculator/-/issues/310