search-provider: omit Gtk.init() and pass argv
We don't need to init GTK as part of the service as nothing from Gtk is
being used. That adds about .5% CPU usage on startup to connect to the
display server, parse CSS, etc. Since this is spawned from interactive
typing by the Shell, avoiding that is useful.

Also pass argv to the app so that `--gapplication-service` and `--help`
are functional.