search-provider: use search-provider as subprocess for isolation
Instead of using the gnome-calculator process for a subprocess, use the
gnome-calculator-search-provider as a subprocess to do isolated search.

Some isolation is necessary because GCalc cannot prevent itself from the
hautling problem apparently. While that needs someone to fix it, this gets
things protected by default.

Additionally, this does the settings lookup from the parent process so
that the child process does not need to load GSettings infrastructure in
the (hopefully) short-lived solver.

Currently, almost all of the overhead in the subprocess is within the
dynamic loader, which is busy loading dynamic libraries. The obvious
way to work around that would be a proxy process which allows a single
process to be used, but can reap it if a timeout expires.