math-buttons: Add insert-brackets
This allows inserting pair of opening and closing characters, either by
enclosing the selection in them or by guessing the right one to insert
at the cursor like Andoid's calculator does for parentheses. For the
moment it does so only for the absolute value which isn't that helpful
as it uses the same opening and closing brackets, but it could be used
later for others bracket pairs like the parentheses, the floor brackets,
and the ceiling brackets.