math-buttons: Insert space before and after text
This inserts a space before an alphabetical symbol if it is preceded by
another alphabetical character. This is similar to what is done when
inserting "squared" with superscript numbers. This also adds a space
after these symbols so we can more easily chain them up without a
keyboard or parentheses. This also drops the now unneeded hardcoded
spaces around the modulus operator.