Make history clickable again
This reverts #295, until a better solution can be found that doesn't break
being able to copy from history with a click.

Prior to 49 clicking an equation or answer in the history would copy it to
the input. In 49 both were made selectable so could be partially selected,
then copied through context menu / middle-click paste (Ctrl+C/Ctrl+V didn't
work).